api/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.24.0
httpx>=0.27.0
google-genai>=1.0.0
python-dotenv>=1.0.0
